+++ goodwatch-webapp/app/ui/Videos.tsx
@@ -63,8 +63,6 @@ export default function Videos({ videos }: VideosProps) {
 		},
 	}
 
-	console.log({ selectedVideos })
-
 	return (
 		<div className="mt-8">
 			<h2 className="text-2xl font-bold">Videos</h2>