+++ goodwatch-webapp/app/ui/main/Header.tsx
@@ -24,12 +24,12 @@ import { GlobalLoading } from "~/ui/nav/GlobalLoading"
 
 export default function Header() {
 	const location = useLocation()
-	const isPage = (pathname: string) => location.pathname === pathname
+	const isPage = (pathname: string) => location.pathname.startsWith(pathname)
 
 	const { user, loading } = useUser()
 
 	return (
-		<div className="bg-gray-900 fixed top-0 z-50 w-full">
+		<div className="fixed top-0 z-50 w-full bg-gray-900">
 			<GlobalLoading />
 			<Disclosure as="nav" className="bg-gray-950/35">
 				{({ open }) => (